#249: Add tests to cover usage of custom objects as argument, variable and return type.
